+++ saleor/graphql/account/mutations/authentication/request_password_reset.py
@@ -15,6 +15,7 @@ from ....core.mutations import BaseMutation
 from ....core.types import AccountError
 from ....core.utils import WebhookEventInfo
 from ....plugins.dataloaders import get_plugin_manager_promise
+from ....site.dataloaders import get_site_promise
 
 
 class RequestPasswordReset(BaseMutation):
@@ -50,7 +51,7 @@ class RequestPasswordReset(BaseMutation):
         ]
 
     @classmethod
-    def clean_user(cls, email, redirect_url):
+    def clean_user(cls, email, redirect_url, info: ResolveInfo):
         try:
             validate_storefront_url(redirect_url)
         except ValidationError as error:
@@ -58,6 +59,7 @@ class RequestPasswordReset(BaseMutation):
                 {"redirect_url": error}, code=AccountErrorCode.INVALID.value
             )
 
+        site = get_site_promise(info.context).get()
         user = retrieve_user_by_email(email)
         if not user:
             raise ValidationError(
@@ -69,7 +71,7 @@ class RequestPasswordReset(BaseMutation):
                 }
             )
 
-        if not user.is_active:
+        if not user.can_login(site.settings):
             raise ValidationError(
                 {
                     "email": ValidationError(
@@ -98,7 +100,7 @@ class RequestPasswordReset(BaseMutation):
         email = data["email"]
         redirect_url = data["redirect_url"]
         channel_slug = data.get("channel")
-        user = cls.clean_user(email, redirect_url)
+        user = cls.clean_user(email, redirect_url, info)
 
         if not user.is_staff:
             channel_slug = clean_channel(