+++ saleor/graphql/account/schema.py
@@ -32,6 +32,7 @@ from .mutations.account import (
     ConfirmAccount,
     ConfirmEmailChange,
     RequestEmailChange,
+    SendConfirmationEmail,
 )
 from .mutations.authentication import (
     CreateToken,
@@ -267,6 +268,7 @@ class AccountMutations(graphene.ObjectType):
     external_verify = ExternalVerify.Field()
 
     request_password_reset = RequestPasswordReset.Field()
+    send_confirmation_email = SendConfirmationEmail.Field()
     confirm_account = ConfirmAccount.Field()
     set_password = SetPassword.Field()
     password_change = PasswordChange.Field()